LayoutTests/platform/ios/imported/w3c/web-platform-tests/dom/events/scrolling/scrollend-event-not-fired-after-removing-scroller.tentative-expected.txt


PASS No scrollend is received after removing parent div
PASS No scrollend is received after removing scrolling element
PASS scrollend is received after removing descendant div
